# openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AffiVideo.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Whether the API can reach its database and object storage.",
        "operationId": "GetStatus",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/antiforgery-token": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "The token every state-changing request must send in the X-CSRF-TOKEN header.",
        "operationId": "GetAntiforgeryToken",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AntiforgeryTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/session": {
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Signs a member in with email and password and starts their session.",
        "operationId": "SignIn",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "Who is signed in, and their Organization.",
        "operationId": "GetSession",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Session"
        ],
        "summary": "Ends the session.",
        "operationId": "SignOut",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "The Organization's settings.",
        "operationId": "GetOrganization",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Organizations"
        ],
        "summary": "Changes the Organization's settings. Owners only.",
        "operationId": "UpdateOrganization",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/members": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "The Organization's members, by email.",
        "operationId": "ListMembers",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfMemberResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Adds an Editor to the Organization. Owners only.",
        "operationId": "AddMember",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/organizations/{organizationId}/audit-log": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "The sensitive actions taken in the Organization, newest first. Owners only.",
        "operationId": "ReadAuditLog",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfAuditLogEntryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddMemberRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AntiforgeryTokenResponse": {
        "required": [
          "requestToken"
        ],
        "type": "object",
        "properties": {
          "requestToken": {
            "type": "string"
          }
        }
      },
      "AuditLogEntryResponse": {
        "required": [
          "id",
          "organizationId",
          "actorMemberId",
          "actorEmail",
          "action",
          "subjectId",
          "occurredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "actorMemberId": {
            "type": "string",
            "format": "uuid"
          },
          "actorEmail": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "subjectId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DependencyStatusResponse": {
        "required": [
          "reachable"
        ],
        "type": "object",
        "properties": {
          "reachable": {
            "type": "boolean"
          }
        }
      },
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "MemberResponse": {
        "required": [
          "id",
          "email",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MemberRole"
          }
        }
      },
      "MemberRole": {
        "enum": [
          "Owner",
          "Editor"
        ]
      },
      "OrganizationResponse": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "PagedResponseOfAuditLogEntryResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogEntryResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "PagedResponseOfMemberResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "total": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SessionResponse": {
        "required": [
          "member",
          "organization"
        ],
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/MemberResponse"
          },
          "organization": {
            "$ref": "#/components/schemas/OrganizationResponse"
          }
        }
      },
      "SignInRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "StatusResponse": {
        "required": [
          "database",
          "objectStorage"
        ],
        "type": "object",
        "properties": {
          "database": {
            "$ref": "#/components/schemas/DependencyStatusResponse"
          },
          "objectStorage": {
            "$ref": "#/components/schemas/DependencyStatusResponse"
          }
        }
      },
      "UpdateOrganizationRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Status"
    },
    {
      "name": "Session"
    },
    {
      "name": "Organizations"
    }
  ]
}