# src/AffiVideo.Domain/MemberRole.cs
namespace AffiVideo.Domain;

/// <summary>What a member may do in their Organization. An Editor does all creative work; only an Owner manages members and Organization settings.</summary>
public enum MemberRole
{
    Owner,
    Editor,
}